# pkg/swagger/docs/swagger.yaml
definitions:
...
  emplacc-api_api_v1_dto.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  emplacc-api_api_v1_dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorDetail'
      meta:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ResponseMeta'
    type: object
...
  emplacc-api_api_v1_dto.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  emplacc-api_api_v1_dto.ResponseMeta:
    properties:
      pagination:
        $ref: '#/definitions/emplacc-api_api_v1_dto.Pagination'
      trace_id:
        type: string
    type: object
...
  emplacc-api_api_v1_dto.UserResponse:
    properties:
      data:
        $ref: '#/definitions/emplacc-api_api_v1_dto_response.User'
      meta:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ResponseMeta'
    type: object
  emplacc-api_api_v1_dto.UsersListResponse:
    properties:
      data:
        $ref: '#/definitions/emplacc-api_api_v1_dto_response.UsersPage'
      meta:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ResponseMeta'
    type: object
...
  emplacc-api_api_v1_dto_request.CreateUser:
    properties:
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      profession:
        type: string
      tg_id:
        type: string
      tg_user_id:
        type: integer
    type: object
...
  emplacc-api_api_v1_dto_response.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_login:
        type: string
      last_name:
        type: string
      profession:
        type: string
      tg_id:
        type: string
      tg_user_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  emplacc-api_api_v1_dto_response.UsersPage:
    properties:
      users:
        items:
          $ref: '#/definitions/emplacc-api_api_v1_dto_response.User'
        type: array
    type: object
...
info:
  contact: {}
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.UsersListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: List Users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User creation payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/emplacc-api_api_v1_dto_request.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Create User
      tags:
      - Users
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Get User
      tags:
      - Users
...
swagger: "2.0"